requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.8.1
cachetools==6.2.3
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
distro==1.9.0
Django==5.1.4
djangorestframework==3.16.1
filelock==3.16.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
groq==0.37.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tk==0.1.0
tkinterdnd2==0.4.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.4.0
